fix alias generation
only launch files in state_machines folder contain rosnode kill /state_machine

+++ env-hooks/99.amigo_functions.bash
@@ -7,7 +7,11 @@ install_services_cmd=""
 for launch_file in ${launch_files}; do
     launch_file_name=$(echo ${launch_file##*/} | cut -d"." -f1)
     name=${bringup_name//_/-}-${launch_file_name//_/-}
-    eval "${name}() { export ROBOT_BRINGUP_PATH=${package_path} && rosnode kill /state_machine && roslaunch ${launch_file} \$@; }"
+    if [[ $launch_file == *"state_machines"* ]]; then
+        eval "${name}() { export ROBOT_BRINGUP_PATH=${package_path} && rosnode kill /state_machine && roslaunch ${launch_file} \$@; }"
+    else
+        eval "${name}() { export ROBOT_BRINGUP_PATH=${package_path} && roslaunch ${launch_file} \$@; }"
+    fi
 
     install_services_cmd="$install_services_cmd && echo 'Creating /etc/systemd/user/$name.service' && rosrun ${bringup_name}_bringup systemd_service_from_launch_file ${launch_file} amigo ~/.tue/setup.bash | sudo tee /etc/systemd/user/$name.service"
 done
